// src/KRINT.Frontend/src/app/api/model/provisionResultDto.ts
/**
 * KRINT.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InnerUserPasswordDto } from './innerUserPasswordDto';
import { ProvisionedDatabaseDto } from './provisionedDatabaseDto';


export interface ProvisionResultDto { 
    instance: ProvisionedDatabaseDto;
    databases: Array<string>;
    users: Array<InnerUserPasswordDto>;
}